Modernize CMake script
- Remove usage of deprecated variables
- Split source file list over several lines
- Color compiler diagnostics if color is forced on Cargo side
- Integrate the rerun if changed mechanism
- Fix incoherence between the `local` header and source file

+++ vm/src/memory/locals.cpp
@@ -10,7 +10,7 @@ int64_t Locals::get_q_word(size_t at) const {
 char Locals::get_byte(size_t at) const {
     return get<char>(at);
 }
-uint64_t Locals::get_ref(uint64_t at) const {
+uint64_t Locals::get_ref(size_t at) const {
     return get<uint64_t>(at);
 }
